Removing IAnnotationContext in favor of IAnnotationFileAware
