+++ mindplay/annotations/standard/ParamAnnotation.php
@@ -13,9 +13,9 @@
 
 namespace mindplay\annotations\standard;
 
-use mindplay\annotations\AnnotationContext;
 use mindplay\annotations\AnnotationException;
-use mindplay\annotations\IAnnotationContext;
+use mindplay\annotations\AnnotationFile;
+use mindplay\annotations\IAnnotationFileAware;
 use mindplay\annotations\IAnnotationParser;
 use mindplay\annotations\Annotation;
 
@@ -24,7 +24,7 @@ use mindplay\annotations\Annotation;
  *
  * @usage('method'=>true, 'inherited'=>true, 'multiple'=>true)
  */
-class ParamAnnotation extends Annotation implements IAnnotationParser, IAnnotationContext
+class ParamAnnotation extends Annotation implements IAnnotationParser, IAnnotationFileAware
 {
     /**
      * @var string
@@ -37,11 +37,11 @@ class ParamAnnotation extends Annotation implements IAnnotationParser, IAnnotati
     public $name;
 
     /**
-         * Annotation file.
-         *
-         * @var AnnotationContext
-         */
-        protected $context;
+     * Annotation file.
+     *
+     * @var AnnotationFile
+     */
+    protected $file;
 
     /**
      * Parse the standard PHP-DOC "param" annotation.
@@ -73,11 +73,18 @@ class ParamAnnotation extends Annotation implements IAnnotationParser, IAnnotati
             throw new AnnotationException('ParamAnnotation requires a name property');
         }
 
-        $this->type = $this->context->resolveType($this->type);
+        $this->type = $this->file->resolveType($this->type);
     }
 
-    public function setAnnotationContext(AnnotationContext $context)
+    /**
+     * Provides information about file, that contains this annotation.
+     *
+     * @param AnnotationFile $file Annotation file.
+     *
+     * @return void
+     */
+    public function setAnnotationFile(AnnotationFile $file)
     {
-        $this->context = $context;
+        $this->file = $file;
     }
 }
